moved Designer Constants to own file and added colors
